Added purge date to ALMA
Purge date is set whenever a person is inactive in campus system, but
still active in ALMA. The purge date is set with a buffer time of 6 months
from actual date on.
Added inactiveamum column in alma table to be able to query only new
inactive user.
Also date inactiveamum is (re-)set to null again for all active user.
This step is needed e.g. if student ended with bachelor - was inactive -
but goes on with master - active again.